add sub command 'node' to app console
