+++ cmd/root.go
@@ -4,9 +4,7 @@ import (
 	"fmt"
 	"os"
 
-	"github.com/icstglobal/plasma/network"
-	homedir "github.com/mitchellh/go-homedir"
-	log "github.com/sirupsen/logrus"
+	"github.com/icstglobal/plasma/cmd/node"
 	"github.com/spf13/cobra"
 	"github.com/spf13/viper"
 )
@@ -15,23 +13,14 @@ var cfgFile string
 
 // rootCmd represents the base command when called without any subcommands
 var rootCmd = &cobra.Command{
-	Use:   "plasma",
-	Short: "The plasma application",
-	Long:  `Start the Plasma child chain.`,
-	// Uncomment the following line if your bare application
-	// has an action associated with it:
-	Run: func(cmd *cobra.Command, args []string) {
-		n := new(network.Node)
-		err := n.Start()
-		if err != nil {
-			log.WithError(err).Fatal("node failed to start")
-		}
-	},
+	Use: "plasma",
 }
 
 // Execute adds all child commands to the root command and sets flags appropriately.
 // This is called by main.main(). It only needs to happen once to the rootCmd.
 func Execute() {
+	initSubCommands()
+
 	if err := rootCmd.Execute(); err != nil {
 		fmt.Println(err)
 		os.Exit(1)
@@ -44,7 +33,7 @@ func init() {
 	// Here you will define your flags and configuration settings.
 	// Cobra supports persistent flags, which, if defined here,
 	// will be global for your application.
-	rootCmd.PersistentFlags().StringVar(&cfgFile, "config", "", "config file (default is $HOME/.plasma.yaml)")
+	rootCmd.PersistentFlags().StringVar(&cfgFile, "config", "", "config file (inherited by sub commands)")
 
 	// Cobra also supports local flags, which will only run
 	// when this action is called directly.
@@ -53,40 +42,21 @@ func init() {
 
 // initConfig reads in config file and ENV variables if set.
 func initConfig() {
+	viper.AutomaticEnv() // read in environment variables that match
+
+	// dont search default config, but leave it to the sub command
 	if cfgFile != "" {
 		// Use config file from the flag.
 		viper.SetConfigFile(cfgFile)
-	} else {
-		// Find home directory.
-		home, err := homedir.Dir()
-		if err != nil {
-			fmt.Println(err)
-			os.Exit(1)
+		// If a config file is found, read it in.
+		if err := viper.ReadInConfig(); err == nil {
+			fmt.Println("Using config file:", viper.ConfigFileUsed())
+		} else {
+			fmt.Println("loading config failed:", err)
 		}
-
-		// Search config in home directory with name ".plasma" (without extension).
-		viper.AddConfigPath(home)
-		viper.AddConfigPath(".")
-		viper.SetConfigName(".plasma")
 	}
+}
 
-	viper.AutomaticEnv() // read in environment variables that match
-
-	// If a config file is found, read it in.
-	if err := viper.ReadInConfig(); err == nil {
-		fmt.Println("Using config file:", viper.ConfigFileUsed())
-
-		slvl := viper.GetString("logger.level")
-		if len(slvl) > 0 {
-			if lvl, err := log.ParseLevel(slvl); err == nil {
-				log.SetLevel(lvl)
-				log.WithField("logger.level", slvl).Debug("change log level")
-			} else {
-				log.WithError(err).WithField("logger.level", slvl).Warn("unknown logger level")
-			}
-
-		}
-	} else {
-		fmt.Println("loading config failed:", err)
-	}
+func initSubCommands() {
+	rootCmd.AddCommand(node.NodeCmd)
 }
